Overall set-up for how the scene can look
Today it's been the first day of post-production, and while a lot of work still remains, we now have a rough outline for how the scene is going to look. :smile:

Forward I'll be trying to finish the building code alone with mouseInput - and merge branches to help aid work on UI and animations. :+1: